pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-demos"
version = "0.1.0"
description = "Demos showcasing AI agent tools with Claude"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "google-auth>=2.30.0",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.140.0",
    "sounddevice>=0.5.0",
    "soundfile>=0.12.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=14.0",
    "python-multipart>=0.0.18",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "httpx>=0.28.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/agent_demos"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["src/agent_demos/demos/appointment_booking"]
omit = [
    "src/agent_demos/demos/appointment_booking/main.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
